(improvement)(headless) Add data permission integrating test
