+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/aspect/S2DataPermissionAspect.java
@@ -1,12 +1,24 @@
 package com.tencent.supersonic.headless.server.aspect;
 
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
+import com.google.common.collect.Sets;
 import com.tencent.supersonic.auth.api.authentication.pojo.User;
+import com.tencent.supersonic.auth.api.authorization.pojo.AuthRes;
+import com.tencent.supersonic.auth.api.authorization.pojo.AuthResGrp;
+import com.tencent.supersonic.auth.api.authorization.pojo.DimensionFilter;
+import com.tencent.supersonic.auth.api.authorization.request.QueryAuthResReq;
 import com.tencent.supersonic.auth.api.authorization.response.AuthorizedResourceResp;
+import com.tencent.supersonic.auth.api.authorization.service.AuthService;
 import com.tencent.supersonic.common.pojo.Constants;
 import com.tencent.supersonic.common.pojo.Filter;
+import com.tencent.supersonic.common.pojo.QueryAuthorization;
+import com.tencent.supersonic.common.pojo.QueryColumn;
+import com.tencent.supersonic.common.pojo.enums.AuthType;
 import com.tencent.supersonic.common.pojo.enums.FilterOperatorEnum;
+import com.tencent.supersonic.common.pojo.enums.SensitiveLevelEnum;
 import com.tencent.supersonic.common.pojo.exception.InvalidArgumentException;
 import com.tencent.supersonic.common.pojo.exception.InvalidPermissionException;
 import com.tencent.supersonic.common.util.jsqlparser.SqlAddHelper;
@@ -18,11 +30,13 @@ import com.tencent.supersonic.headless.api.pojo.response.DimensionResp;
 import com.tencent.supersonic.headless.api.pojo.response.ModelResp;
 import com.tencent.supersonic.headless.api.pojo.response.SemanticQueryResp;
 import com.tencent.supersonic.headless.api.pojo.response.SemanticSchemaResp;
+import com.tencent.supersonic.headless.api.pojo.response.ViewResp;
 import com.tencent.supersonic.headless.server.pojo.MetaFilter;
 import com.tencent.supersonic.headless.server.pojo.ModelFilter;
 import com.tencent.supersonic.headless.server.service.DimensionService;
 import com.tencent.supersonic.headless.server.service.ModelService;
 import com.tencent.supersonic.headless.server.service.SchemaService;
+import com.tencent.supersonic.headless.server.service.ViewService;
 import com.tencent.supersonic.headless.server.utils.QueryStructUtils;
 import lombok.extern.slf4j.Slf4j;
 import net.sf.jsqlparser.JSQLParserException;
@@ -33,18 +47,22 @@ import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Pointcut;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 import org.springframework.util.CollectionUtils;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.StringJoiner;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import static com.tencent.supersonic.common.pojo.Constants.MINUS;
 
@@ -52,7 +70,10 @@ import static com.tencent.supersonic.common.pojo.Constants.MINUS;
 @Aspect
 @Order(1)
 @Slf4j
-public class S2DataPermissionAspect extends AuthCheckBaseAspect {
+public class S2DataPermissionAspect {
+
+    private static final ObjectMapper MAPPER = new ObjectMapper().setDateFormat(
+            new SimpleDateFormat(Constants.DAY_FORMAT));
 
     @Autowired
     private QueryStructUtils queryStructUtils;
@@ -60,10 +81,12 @@ public class S2DataPermissionAspect extends AuthCheckBaseAspect {
     private DimensionService dimensionService;
     @Autowired
     private ModelService modelService;
-    @Value("${permission.data.enable:true}")
-    private Boolean permissionDataEnable;
     @Autowired
     private SchemaService schemaService;
+    @Autowired
+    private ViewService viewService;
+    @Autowired
+    private AuthService authService;
 
     @Pointcut("@annotation(com.tencent.supersonic.headless.server.annotation.S2DataPermission)")
     private void s2PermissionCheck() {
@@ -71,11 +94,6 @@ public class S2DataPermissionAspect extends AuthCheckBaseAspect {
 
     @Around("s2PermissionCheck()")
     public Object doAround(ProceedingJoinPoint joinPoint) throws Throwable {
-        log.info("s2 permission check!");
-        if (!permissionDataEnable) {
-            log.info("not to check permission!");
-            return joinPoint.proceed();
-        }
         Object[] objects = joinPoint.getArgs();
         SemanticQueryReq queryReq = (SemanticQueryReq) objects[0];
         if (!queryReq.isNeedAuth()) {
@@ -86,13 +104,14 @@ public class S2DataPermissionAspect extends AuthCheckBaseAspect {
         if (Objects.isNull(user) || Strings.isNullOrEmpty(user.getName())) {
             throw new RuntimeException("please provide user information");
         }
+        List<Long> modelIds = getModelsInView(queryReq);
 
         // determine whether admin of the model
-        if (doModelAdmin(user, queryReq.getModelIds())) {
+        if (doModelAdmin(user, modelIds)) {
             return joinPoint.proceed();
         }
         // determine whether the subject field is visible
-        doModelVisible(user, queryReq.getModelIds());
+        doModelVisible(user, modelIds);
 
         if (queryReq instanceof QuerySqlReq) {
             return checkSqlPermission(joinPoint, (QuerySqlReq) queryReq);
@@ -225,7 +244,7 @@ public class S2DataPermissionAspect extends AuthCheckBaseAspect {
         AuthorizedResourceResp authorizedResource = getAuthorizedResource(user,
                 modelIdInView, sensitiveResReq);
         // get sensitiveRes that user has privilege
-        Set<String> resAuthSet = getAuthResNameSet(authorizedResource, queryStructReq.getModelIds());
+        Set<String> resAuthSet = getAuthResNameSet(authorizedResource, modelIdInView);
 
         // if sensitive fields without permission are involved in filter, thrown an exception
         doFilterCheckLogic(queryStructReq, resAuthSet, sensitiveResReq);
@@ -324,6 +343,227 @@ public class S2DataPermissionAspect extends AuthCheckBaseAspect {
             log.info("after doRowPermission, queryStructReq:{}", queryStructReq);
         }
 
+    }
+
+    public boolean doModelAdmin(User user, List<Long> modelIds) {
+        List<ModelResp> modelListAdmin = modelService.getModelListWithAuth(user, null, AuthType.ADMIN);
+        if (CollectionUtils.isEmpty(modelListAdmin)) {
+            return false;
+        } else {
+            Set<Long> modelAdmins = modelListAdmin.stream().map(ModelResp::getId).collect(Collectors.toSet());
+            return !CollectionUtils.isEmpty(modelAdmins) && modelAdmins.containsAll(modelIds);
+        }
+    }
+
+    public void doModelVisible(User user, List<Long> modelIds) {
+        List<Long> modelListVisible = modelService.getModelListWithAuth(user, null, AuthType.VISIBLE)
+                .stream().map(ModelResp::getId).collect(Collectors.toList());
+        modelIds.removeAll(modelListVisible);
+        if (!CollectionUtils.isEmpty(modelIds)) {
+            MetaFilter metaFilter = new MetaFilter();
+            metaFilter.setIds(modelIds);
+            List<ModelResp> modelResps = modelService.getModelList(metaFilter);
+            ModelResp modelResp = modelResps.stream().findFirst().orElse(null);
+            if (modelResp == null) {
+                throw new InvalidArgumentException("查询的模型不存在");
+            }
+            String message = String.format("您没有模型[%s]权限，请联系管理员%s开通", modelResp.getName(), modelResp.getAdmins());
+            throw new InvalidPermissionException(message);
+        }
+    }
+
+    public Set<String> getHighSensitiveColsByModelId(SemanticSchemaResp semanticSchemaResp) {
+        Set<String> highSensitiveCols = new HashSet<>();
+        if (!CollectionUtils.isEmpty(semanticSchemaResp.getDimensions())) {
+            semanticSchemaResp.getDimensions().stream().filter(dimSchemaResp ->
+                            SensitiveLevelEnum.HIGH.getCode().equals(dimSchemaResp.getSensitiveLevel()))
+                    .forEach(dim -> highSensitiveCols.add(dim.getBizName()));
+        }
+        if (!CollectionUtils.isEmpty(semanticSchemaResp.getMetrics())) {
+            semanticSchemaResp.getMetrics().stream().filter(metricSchemaResp ->
+                            SensitiveLevelEnum.HIGH.getCode().equals(metricSchemaResp.getSensitiveLevel()))
+                    .forEach(metric -> highSensitiveCols.add(metric.getBizName()));
+        }
+        return highSensitiveCols;
+    }
+
+    public AuthorizedResourceResp getAuthorizedResource(User user, List<Long> modelIds,
+                                                        Set<String> sensitiveResReq) {
+        List<AuthRes> resourceReqList = new ArrayList<>();
+        sensitiveResReq.forEach(res -> resourceReqList.add(new AuthRes(modelIds.get(0), res)));
+        QueryAuthResReq queryAuthResReq = new QueryAuthResReq();
+        queryAuthResReq.setResources(resourceReqList);
+        queryAuthResReq.setModelIds(modelIds);
+        AuthorizedResourceResp authorizedResource = fetchAuthRes(queryAuthResReq, user);
+        log.info("user:{}, domainId:{}, after queryAuthorizedResources:{}", user.getName(), modelIds,
+                authorizedResource);
+        return authorizedResource;
+    }
+
+    private AuthorizedResourceResp fetchAuthRes(QueryAuthResReq queryAuthResReq, User user) {
+        log.info("queryAuthResReq:{}", queryAuthResReq);
+        return authService.queryAuthorizedResources(queryAuthResReq, user);
+    }
+
+    public Set<String> getAuthResNameSet(AuthorizedResourceResp authorizedResource, List<Long> modelIds) {
+        Set<String> resAuthName = new HashSet<>();
+        List<AuthResGrp> authResGrpList = authorizedResource.getResources();
+        authResGrpList.stream().forEach(authResGrp -> {
+            List<AuthRes> cols = authResGrp.getGroup();
+            if (!CollectionUtils.isEmpty(cols)) {
+                cols.stream().filter(col -> modelIds.contains(col.getModelId()))
+                        .forEach(col -> resAuthName.add(col.getName()));
+            }
+
+        });
+        log.info("resAuthName:{}", resAuthName);
+        return resAuthName;
+    }
+
+    public SemanticQueryResp getQueryResultWithColumns(SemanticQueryResp resultWithColumns,
+                                                       List<Long> modelIds,
+                                                       AuthorizedResourceResp authResource) {
+        addPromptInfoInfo(modelIds, resultWithColumns, authResource, Sets.newHashSet());
+        return resultWithColumns;
+    }
+
+    public SemanticQueryResp desensitizationData(SemanticQueryResp raw, Set<String> need2Apply) {
+        log.debug("start desensitizationData logic");
+        if (CollectionUtils.isEmpty(need2Apply)) {
+            log.info("user has all sensitiveRes");
+            return raw;
+        }
+
+        List<QueryColumn> columns = raw.getColumns();
+
+        boolean doDesensitization = false;
+        for (QueryColumn queryColumn : columns) {
+            for (String sensitiveCol : need2Apply) {
+                if (queryColumn.getNameEn().contains(sensitiveCol)) {
+                    doDesensitization = true;
+                    break;
+                }
+            }
+        }
+        if (!doDesensitization) {
+            return raw;
+        }
+
+        SemanticQueryResp queryResultWithColumns = raw;
+        try {
+            queryResultWithColumns = deepCopyResult(raw);
+        } catch (Exception e) {
+            log.warn("deepCopyResult: ", e);
+        }
+        addAuthorizedSchemaInfo(queryResultWithColumns.getColumns(), need2Apply);
+        desensitizationInternal(queryResultWithColumns.getResultList(), need2Apply);
+        return queryResultWithColumns;
+    }
+
+    private void addAuthorizedSchemaInfo(List<QueryColumn> columns, Set<String> need2Apply) {
+        if (CollectionUtils.isEmpty(need2Apply)) {
+            return;
+        }
+        columns.stream().forEach(col -> {
+            if (need2Apply.contains(getName(col.getNameEn()))) {
+                col.setAuthorized(false);
+            }
+        });
+    }
+
+    private String getName(String nameEn) {
+        Pattern pattern = Pattern.compile("\\((.*?)\\)");
+        Matcher matcher = pattern.matcher(nameEn);
+        if (matcher.find()) {
+            return matcher.group(1).replaceAll("`", "");
+        }
+        return nameEn;
+    }
+
+    private void desensitizationInternal(List<Map<String, Object>> result, Set<String> need2Apply) {
+        log.info("start desensitizationInternal logic");
+        for (int i = 0; i < result.size(); i++) {
+            Map<String, Object> row = result.get(i);
+            Map<String, Object> newRow = new HashMap<>();
+            for (String col : row.keySet()) {
+                boolean sensitive = false;
+                for (String sensitiveCol : need2Apply) {
+                    if (col.contains(sensitiveCol)) {
+                        sensitive = true;
+                        break;
+                    }
+                }
+                if (sensitive) {
+                    newRow.put(col, "******");
+                } else {
+                    newRow.put(col, row.get(col));
+                }
+            }
+            result.set(i, newRow);
+        }
+    }
+
+    private SemanticQueryResp deepCopyResult(SemanticQueryResp raw) throws Exception {
+        SemanticQueryResp queryResultWithColumns = new SemanticQueryResp();
+        BeanUtils.copyProperties(raw, queryResultWithColumns);
+
+        List<QueryColumn> columns = new ArrayList<>();
+        if (!CollectionUtils.isEmpty(raw.getColumns())) {
+            String columnsStr = MAPPER.writeValueAsString(raw.getColumns());
+            columns = MAPPER.readValue(columnsStr, new TypeReference<List<QueryColumn>>() {
+            });
+            queryResultWithColumns.setColumns(columns);
+        }
+        queryResultWithColumns.setColumns(columns);
+
+        List<Map<String, Object>> resultData = new ArrayList<>();
+        if (!CollectionUtils.isEmpty(raw.getResultList())) {
+            for (Map<String, Object> line : raw.getResultList()) {
+                Map<String, Object> newLine = new HashMap<>();
+                newLine.putAll(line);
+                resultData.add(newLine);
+            }
+        }
+        queryResultWithColumns.setResultList(resultData);
+        return queryResultWithColumns;
+    }
+
+    public void addPromptInfoInfo(List<Long> modelIds, SemanticQueryResp queryResultWithColumns,
+                                  AuthorizedResourceResp authorizedResource, Set<String> need2Apply) {
+        List<DimensionFilter> filters = authorizedResource.getFilters();
+        if (CollectionUtils.isEmpty(need2Apply) && CollectionUtils.isEmpty(filters)) {
+            return;
+        }
+        List<String> admins = modelService.getModelAdmin(modelIds.get(0));
+        if (!CollectionUtils.isEmpty(need2Apply)) {
+            String promptInfo = String.format("当前结果已经过脱敏处理， 申请权限请联系管理员%s", admins);
+            queryResultWithColumns.setQueryAuthorization(new QueryAuthorization(promptInfo));
+        }
+        if (!CollectionUtils.isEmpty(filters)) {
+            log.debug("dimensionFilters:{}", filters);
+            ModelResp modelResp = modelService.getModel(modelIds.get(0));
+            List<String> exprList = new ArrayList<>();
+            List<String> descList = new ArrayList<>();
+            filters.stream().forEach(filter -> {
+                descList.add(filter.getDescription());
+                exprList.add(filter.getExpressions().toString());
+            });
+            String promptInfo = "当前结果已经过行权限过滤，详细过滤条件如下:%s, 申请权限请联系管理员%s";
+            String message = String.format(promptInfo, CollectionUtils.isEmpty(descList) ? exprList : descList, admins);
+
+            queryResultWithColumns.setQueryAuthorization(
+                    new QueryAuthorization(modelResp.getName(), exprList, descList, message));
+            log.info("queryResultWithColumns:{}", queryResultWithColumns);
+        }
+    }
+
+    private List<Long> getModelsInView(SemanticQueryReq queryReq) {
+        List<Long> modelIds = queryReq.getModelIds();
+        if (queryReq.getViewId() != null) {
+            ViewResp viewResp = viewService.getView(queryReq.getViewId());
+            modelIds = viewResp.getAllModels();
+        }
+        return modelIds;
     }
 
 }
